+++ src/main/java/com/marginallyclever/ro3/apps/pathtracer/spatialaccelerationstructure/BVHNode.java
@@ -1,5 +1,6 @@
-package com.marginallyclever.ro3.apps.pathtracer;
+package com.marginallyclever.ro3.apps.pathtracer.spatialaccelerationstructure;
 
+import com.marginallyclever.ro3.apps.pathtracer.PathTriangle;
 import com.marginallyclever.ro3.mesh.AABB;
 
 import java.util.List;
