+++ src/test/java/com/marginallyclever/ro3/apps/pathtracer/PathTracerTest.java
@@ -1,10 +1,9 @@
 package com.marginallyclever.ro3.apps.pathtracer;
 
+import com.marginallyclever.ro3.apps.pathtracer.halton.HaltonWithMemory;
 import org.junit.jupiter.api.Test;
 
 import javax.vecmath.Vector3d;
-import java.util.Random;
-import java.util.SplittableRandom;
 
 public class PathTracerTest {
     @Test
